requirements-lock.txt
aniso8601==7.0.0
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coverage==5.0.4
docker==4.2.0
Flask==1.1.2
Flask-GraphQL==2.0.1
fxq-commons==1.1.0
gitdb==4.0.2
GitPython==3.1.0
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.3
MarkupSafe==1.1.1
more-itertools==8.2.0
multidispatch==0.2
multipledispatch==0.6.0
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
promise==2.3
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
Rx==1.6.1
six==1.14.0
smmap==3.0.1
stringcase==1.2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
